refactor: move commands and highlights to plugin file
Move commands, highlight definitions and treesitter registration from
init.lua to a proper plugin/CopilotChat.lua file. This follows better
plugin organization practices and ensures proper loading sequence.

Includes:
- Moving highlight definitions
- Moving commands (CopilotChat*)
- Moving autocmd for directory tracking
- Moving treesitter language registration
- Adding proper plugin loading guard
